feat(controller): add validation to hotels
add authentication and authorization validation to the next methods

- createHotelHandler
- updateHotelHandler
- deleteHotelHandler

+++ src/main/java/com/revature/controller/HotelController.java
@@ -2,8 +2,11 @@ package com.revature.controller;
 
 
 import com.revature.dto.request.HotelFilterDTO;
+import com.revature.exceptions.ForbiddenActionException;
 import com.revature.exceptions.ResourceNotFoundException;
+import com.revature.exceptions.UnauthenticatedException;
 import com.revature.models.Hotel;
+import com.revature.models.Role;
 import com.revature.services.HotelService;
 import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -28,16 +31,16 @@ public class HotelController {
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
     public Hotel createHotelHandler(@RequestBody Hotel hotel, HttpSession session){
-//        if (session.getAttribute("userId") == null){
-//            throw new UnauthenticatedException("User is not authenticated");
-//        }
-//
-//        // I need to make sure the role of the user is a teacher
-//        if (session.getAttribute("role") != Role.OWNER){
-//            throw new ForbiddenActionException("You must be a teacher to access this");
-//        }
+        if (session.getAttribute("userId") == null){
+            throw new UnauthenticatedException("User is not authenticated");
+        }
 
-        return hotelService.createHotel(hotel);
+        // I need to make sure the role of the user is a teacher
+        if (session.getAttribute("role") != Role.OWNER){
+            throw new ForbiddenActionException("You must be a teacher to access this");
+        }
+
+        return hotelService.createHotel((int) session.getAttribute("userId"), hotel);
     }
 
     @GetMapping
@@ -58,16 +61,20 @@ public class HotelController {
 
     @PutMapping("{hotelId}")
     public Optional<Hotel> updateHotelHandler(@PathVariable int hotelId, @RequestBody Hotel updatedHotel, HttpSession session) {
-//        if (session.getAttribute("userId") == null) {
-//            throw new UnauthenticatedException("User is not authenticated");
-//        }
-//
-//        if (session.getAttribute("role") != Role.OWNER) {
-//            throw new ForbiddenActionException("You must be an owner to modify this hotel");
-//        }
+        if (session.getAttribute("userId") == null) {
+            throw new UnauthenticatedException("User is not authenticated");
+        }
+
+        if (session.getAttribute("role") != Role.OWNER) {
+            throw new ForbiddenActionException("You must be an owner to modify this hotel");
+        }
 
         if(hotelService.checkHotelExisting(hotelId)){
             throw new ResourceNotFoundException("No hotel with id: " + hotelId);
+        }
+
+        if (hotelService.isUserOwnerOfHotel((int) session.getAttribute("userId"),hotelId)){
+            throw new ForbiddenActionException("You must be the owner of this hotel to make updates");
         }
 
         return hotelService.updateHotel(hotelId, updatedHotel);
@@ -76,16 +83,20 @@ public class HotelController {
     @DeleteMapping("{hotelId}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
     public void deleteHotelHandler(@PathVariable int hotelId, HttpSession session) {
-//        if (session.getAttribute("userId") == null) {
-//            throw new UnauthenticatedException("User is not authenticated");
-//        }
-//
-//        if (session.getAttribute("role") != Role.OWNER) {
-//            throw new ForbiddenActionException("You must be an owner to delete this hotel");
-//        }
+        if (session.getAttribute("userId") == null) {
+            throw new UnauthenticatedException("User is not authenticated");
+        }
+
+        if (session.getAttribute("role") != Role.OWNER) {
+            throw new ForbiddenActionException("You must be an owner to modify this hotel");
+        }
 
         if(hotelService.checkHotelExisting(hotelId)){
             throw new ResourceNotFoundException("No hotel with id: " + hotelId);
+        }
+
+        if (hotelService.isUserOwnerOfHotel((int) session.getAttribute("userId"),hotelId)){
+            throw new ForbiddenActionException("You must be the owner of this hotel to make updates");
         }
 
         hotelService.deleteHotel(hotelId);
